pyproject.toml

[tool.uv]
index = [
    { name = "pytorch-nightly", url = "https://download.pytorch.org/whl/nightly/cu128", explicit = true },
]
override-dependencies = [
    "torch>=2.7.0.dev",
    "torchaudio>=2.7.0.dev",
    "torchvision>=0.22.0.dev",
]

[tool.uv.sources]
torch = { index = "pytorch-nightly" }
torchaudio = { index = "pytorch-nightly" }
torchvision = { index = "pytorch-nightly" }

[project]
name = "chatterbox-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.11.*"
dependencies = [
    "chatterbox-tts>=0.1.6",
    "fastapi>=0.128.0",
    "hf-xet>=1.2.0",
    "ml-dtypes>=0.4.0",
    "onnx>=1.19.0",
    "resemble-perth>=1.0.1",
    "uvicorn>=0.40.0",
    "torch",
    "torchaudio",
]
